fixed multiple styles and finally created correct fresh_token
