+++ package.json
@@ -19,10 +19,10 @@
     "@example/basics": "link:",
     "@types/node": "^18.11.17",
     "astro": "^1.6.8",
+    "astro-icon": "^0.8.0",
     "mdast-util-to-string": "^3.1.0",
     "motion": "^10.14.3",
     "reading-time": "^1.5.0",
-    "tailwindcss": "^3.0.24",
-    "ts-node": "^10.9.1"
+    "tailwindcss": "^3.0.24"
   }
 }
