+++ src/components/Navbar.astro
@@ -6,9 +6,9 @@ import SpotifyNowPlaying from "./SpotifyNowPlaying.astro";
   role="navigation"
   aria-label="Main"
   id="top-nav"
-  class="flex flex-col md:flex-row gap-4 justify-center items-center w-full pt-8 pb-8 sm:pb-16"
+  class="flex flex-col md:flex-row md:justify-evenly gap-4 items-center pt-8 pb-8 sm:pb-16"
 >
-  <ul class="flex justify-center gap-4 mr-0 md:mr-[350px]">
+  <ul class="flex justify-center gap-4">
     <a
       href="/"
       class="relative font-medium text-[#87CB49] before:absolute before:-bottom-1 before:h-0.5 before:w-full before:scale-x-0 before:bg-[#87CB49] before:transition hover:before:scale-x-100"
