+++ src/components/TechBadge.astro
@@ -14,5 +14,5 @@ const { name, imgSrc, altSrc } = Astro.props;
   >
     <img src={imgSrc} alt={altSrc} height="30" width="30" />
   </div>
-  <h3>{name}</h3>
+  <h3 class="text-[14px] md:text-[16px]">{name}</h3>
 </div>
