+++ src/env.d.ts
@@ -4,7 +4,6 @@ interface ImportMetaEnv {
     readonly SPOTIFY_REFRESH_TOKEN: string;
     readonly SPOTIFY_CLIENT_ID: string;
     readonly SPOTIFY_CLIENT_SECRET: string;
-    readonly URL: string;
 }
 
 interface ImportMeta {
