+++ src/lib/spotify.ts
@@ -5,31 +5,42 @@ type Artist = {
     followers: { total: number },
 }
 
-export const getAccessToken = async () => {
-    let refresh_token = import.meta.env.SPOTIFY_REFRESH_TOKEN;
-    let clientId = import.meta.env.SPOTIFY_CLIENT_ID;
-    let clientSecret = import.meta.env.SPOTIFY_CLIENT_SECRET;
+export type NowPlayingTrackResponse = {
+    isPlaying: boolean;
+    title: string;
+    artist: string;
+    url: string;
+    img: string;
+};
 
-    const response = await fetch('https://accounts.spotify.com/api/token', {
+
+// Get access token from Spotify
+export const getAccessToken = async () => {
+    // Get environment variables
+    const refresh_token = import.meta.env.SPOTIFY_REFRESH_TOKEN;
+    const client_id = import.meta.env.SPOTIFY_CLIENT_ID;
+    const client_secret = import.meta.env.SPOTIFY_CLIENT_SECRET;
+
+    const basic = Buffer.from(`${client_id}:${client_secret}`).toString("base64")
+    const TOKEN_ENDPOINT = "https://accounts.spotify.com/api/token";
+
+    const response = await fetch(TOKEN_ENDPOINT, {
         method: "POST",
         headers: {
-            Authorization: `Basic ${Buffer.from(
-                `${clientId}:${clientSecret}`
-            ).toString("base64")}`,
+            Authorization: `Basic ${basic}`,
             "Content-Type": "application/x-www-form-urlencoded",
         },
         body: new URLSearchParams({
             grant_type: "refresh_token",
             refresh_token,
         }),
-    })
-
+    });
     return response.json();
 }
 
+// Get authorization for currently playing scope
 const nowPlaying = async () => {
     const { access_token } = await getAccessToken();
-
     return fetch("https://api.spotify.com/v1/me/player/currently-playing", {
         headers: {
             Authorization: `Bearer ${access_token}`,
@@ -37,7 +48,29 @@ const nowPlaying = async () => {
     });
 };
 
-console.log(getAccessToken())
+//Get currently playing track
+export const get = async () => {
+    const response = await nowPlaying();
 
-export { nowPlaying };
-export type { Artist };
+    if (response.status === 204 || response.status > 400) {
+        return new Response(JSON.stringify({ isPlaying: false }), {
+            status: 200,
+        });
+    }
+
+    const { item } = await response.json();
+
+    const track = {
+        isPlaying: true,
+        title: item.name,
+        artist: item.artists
+            .map((_artist: Artist) => _artist.name)
+            .join(", "),
+        url: item.external_urls.spotify,
+        img: item.album.images[0].url,
+    };
+
+    return new Response(JSON.stringify(track), {
+        status: 200,
+    });
+};
