+++ src/pages/index.astro
@@ -70,7 +70,7 @@ const repos = await pinnedRepos.json();
       </div>
       <img
         id="pfp-1"
-        class="mb-10 md:mb-0 h-[120px] w-[120px] md:h-[240px] md:w-[240px] rounded-full ml-10"
+        class="mb-10 md:mb-0 h-[120px] w-[120px] md:h-[240px] md:w-[240px] rounded-full ml-0 md:ml-10"
         src="https://avatars.githubusercontent.com/u/85648028?v=4"
         alt="Paul Valladares"
       />
